test: split str-Enum downcast Visibility coverage into its own file
tests/test_lambda_identity_ordinal.py sits at the AI-readiness 1200-line
test-file cap with zero headroom; the TestStrSubclassEnumsSurviveTheWalk
Unrewritten class added in 49050e0 pushed it to 1372 lines, tripping
scripts/check_architecture.py's new-test-size gate. Moved that class (plus
its _walk_iterable_values helper) verbatim to a new
tests/test_str_enum_downcast_walk.py, matching this repo's existing split
convention (test_raw_pre_strip_snapshot_normalization.py out of the same
parent file). Updated the serialization.str_enum_downcast_via_generic_rewrite
bug-class entry's seed_tests to the new path. No test logic changed;
tests/test_lambda_identity_ordinal.py is back to byte-identical with its
pre-49050e0 content.
